pyproject.toml
[project]
name = "luca"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.123.5",
    "google-adk>=1.18.0",
    "google-cloud-firestore>=2.21.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "sqlmodel>=0.0.27",
]
